IndexedDB/idbobjectstore_createIndex2.htm

<!DOCTYPE html>
<html>
<head>
    <meta content="text/html; charset=utf-8" http-equiv="content-type" />
    <title> IDBObjectStore.createIndex() - attempt to create an index that requires unique values on an object store already contains duplicates </title>
    <script type="text/javascript" src="support.js"></script>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script type="text/javascript">
        var objectStoreName = "objectstore";
        var record = {indexedProperty:"bar"};
        var db = null;
        
        var t = async_test();
        
        function RunTest() {
            var rqOpen = window.indexedDB.open(databaseName, databaseVersion);
            rqOpen.onupgradeneeded = t.step_func(function(event) {
                db = event.target.result;
                db.onerror = t.step_func(assert_database_error);
                var objStore = db.createObjectStore(objectStoreName, {autoIncrement:true});
                objStore.add(record);
                objStore.add(record);
                var index = objStore.createIndex("i1", "indexedProperty", {unique:true});
                assert_true(index instanceof IDBIndex);
                event.target.transaction.onabort = t.step_func(function(event) { t.done(); });
                event.target.transaction.oncomplete = t.step_func(assert_unexpected_complete);
            });
            rqOpen.onerror = t.step_func(assert_open_request_error);
        }
        
        add_completion_callback(function() { if(db) db.close(); });
        
        setup(function() {
            var rqDelete = window.indexedDB.deleteDatabase(databaseName);
            rqDelete.onsuccess = t.step_func( RunTest );
        });
    </script>
</head>
<body>
    <div id="log"></div>
</body>
</html>
